make GetLatestPushedSha async
