+++ src/GitHub.Exports/Models/ISimpleRepositoryModel.cs
@@ -1,6 +1,7 @@
 ﻿using GitHub.Primitives;
 using GitHub.UI;
 using System.ComponentModel;
+using System.Threading.Tasks;
 
 namespace GitHub.Models
 {
@@ -20,6 +21,6 @@ namespace GitHub.Models
         /// Updates the url information based on the local path
         /// </summary>
         void Refresh();
-        UriString GenerateUrl(string path = null, int startLine = -1, int endLine = -1);
+        Task<UriString> GenerateUrl(string path = null, int startLine = -1, int endLine = -1);
     }
 }
