+++ src/GitHub.Exports/Models/SimpleRepositoryModel.cs
@@ -7,6 +7,7 @@ using GitHub.Primitives;
 using GitHub.UI;
 using GitHub.VisualStudio.Helpers;
 using GitHub.Services;
+using System.Threading.Tasks;
 
 namespace GitHub.Models
 {
@@ -72,12 +73,13 @@ namespace GitHub.Models
         /// <param name="startLine">A specific line, or (if specifying the <paramref name="endLine"/> as well) the start of a range</param>
         /// <param name="endLine">The end of a line range on the specified file.</param>
         /// <returns>An UriString with the generated url, or null if the repository has no remote server configured or if it can't be found locally</returns>
-        public UriString GenerateUrl(string path = null, int startLine = -1, int endLine = -1)
+        public async Task<UriString> GenerateUrl(string path = null, int startLine = -1, int endLine = -1)
         {
             if (CloneUrl == null)
                 return null;
 
-            var sha = GetLatestPushedSha();
+            var sha = await GitService.GitServiceHelper.GetLatestPushedSha(LocalPath);
+            
             // this also incidentally checks whether the repo has a valid LocalPath
             if (String.IsNullOrEmpty(sha))
                 return CloneUrl.ToRepositoryUrl().AbsoluteUri;
@@ -109,22 +111,6 @@ namespace GitHub.Models
             }
 
             return new UriString(GenerateUrl(CloneUrl.ToRepositoryUrl().AbsoluteUri, sha, path, startLine, endLine));
-        }
-
-        string GetLatestPushedSha()
-        {
-            var repo = GitService.GitServiceHelper.GetRepository(LocalPath);
-            var remoteHeads = repo.Refs.Where(r => r.IsRemoteTrackingBranch).ToList();
-
-            foreach (var c in repo.Commits)
-            {
-                if (repo.Refs.ReachableFrom(remoteHeads, new[] { c }).Any())
-                {
-                    return c.Sha;
-                }
-            }
-
-            return null;
         }
 
         const string CommitFormat = "{0}/commit/{1}";
