+++ src/GitHub.VisualStudio/Menus/CopyLink.cs
@@ -33,7 +33,7 @@ namespace GitHub.VisualStudio.Menus
             if (!isgithub)
                 return;
 
-            var link = GenerateLink();
+            var link = await GenerateLink();
             if (link == null)
                 return;
             try
